Change int speed to double speed, fix errors and write unit tests for update()-method in FootSoldier class.

+++ src/enemy.hpp
@@ -4,12 +4,12 @@
 
 class Enemy: public GameObject {
 protected:
-    int speed_;
+    double speed_;
     int value_;
     int price_;
     int maxHP_;
 public:
-    explicit Enemy(int speed, int value, int price, int maxHP, sf::Sprite sprite, Game& game, Pos position)
+    explicit Enemy(double speed, int value, int price, int maxHP, sf::Sprite sprite, Game& game, Pos position)
             : GameObject(sprite, game, position), speed_(speed), value_(value), price_(price), maxHP_(maxHP) {};
     
     virtual ~Enemy() {};
